backend/pyproject.toml
[project]
name = "handwrite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "handright==8.1.0",
    "pillow==9.4.0",
]
